Removing some warnings

+++ src/shared/state/Tribe_Stack.cpp
@@ -38,7 +38,7 @@ Tribe* Tribe_Stack::take_tribe_at_position(int position)
 std::vector<Tribe*> Tribe_Stack::get_tribes_on_top()
 {
     std::vector<Tribe*> seeable_Tribes;
-    for (size_t i = 0; i < std::min(6, static_cast<int>(stack.size())); ++i)
+    for (size_t i = 0; i < std::min(static_cast<size_t>(6), stack.size()); ++i)
     {
         seeable_Tribes.push_back(&stack[i]);
     }
